extracted out get_outdir function
