+++ inc/util/cli_utils.hpp
@@ -21,6 +21,7 @@
 #include <vector>
 
 #include <boost/program_options.hpp>
+#include <boost/filesystem/operations.hpp>
 
 #include "util/logger.hpp"
 #include "vcf/validator.hpp"
@@ -45,6 +46,27 @@ namespace ebi
         po::notify(vm);
         return vm;
     }
+
+    std::string get_output_path(const std::string &outdir, const std::string &file_path)
+    {
+        if (outdir == "") {
+            return file_path;
+        }
+
+        boost::filesystem::path file_boost_path{file_path};
+        boost::filesystem::path outdir_boost_path{outdir};
+        if (!boost::filesystem::exists(outdir_boost_path)) {
+            throw std::invalid_argument{"Directory not found: " + outdir_boost_path.string()};
+        }
+        if (!boost::filesystem::is_directory(outdir_boost_path)) {
+            throw std::invalid_argument{"outdir should be a directory, not a file: " + outdir_boost_path.string()};
+        }
+
+        outdir_boost_path /= file_boost_path.filename();
+
+        return outdir_boost_path.string();
+    }
+
   }
 }
